requirements.txt
notion-client
pyyaml
discord